backend/requirements.txt
fastapi
uvicorn
qrcode[pil]
boto3
python-multipart
python-dotenv